+++ functions/resize-images.js
@@ -7,13 +7,12 @@ const sharp = require('sharp')
 const glob = util.promisify(globCb)
 const readFile = util.promisify(fs.readFile)
 
-const { sizes, resizedDir, outputDir } = require('../src/util/getImageUrl')
-const dirPrefix = './public'
+const { sizes, imgixUrl } = require('../src/util/getImageUrl')
 
 const options = {
+  inputDir: './public/images/uploads',
+  outputDir: './public/images/uploads/resized',
   sizes,
-  inputDir: `${dirPrefix}${outputDir}`,
-  outputDir: `${dirPrefix}${resizedDir}`,
   imageFormats: ['jpg', 'jpeg', 'png', 'gif', 'webp']
 }
 
@@ -41,7 +40,7 @@ const saveImages = ({ buffer, filename }) => {
         filename,
         extname
       )}.${size}${extname}`
-      const outputFile = path.resolve(options.outputDir, newFilename)
+      const outputFile = `${options.outputDir}/${newFilename}`
       const fileExists = await doesFileExist({ filename: outputFile })
       if (fileExists) return console.log(`↩️  ${outputFile} exists, skipping`)
       return saveImage({ buffer, size, outputFile })
@@ -87,4 +86,8 @@ const resizeImages = async () => {
   }
 }
 
-resizeImages()
+if (imgixUrl) {
+  console.log(`📡  Using imgix to resize images: ${imgixUrl}`)
+} else {
+  resizeImages()
+}