+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.3 - 2018-05-21
+
+* Add optional imgix integration for images
+* Return to console.log for build console messages
+
+## 0.2.2 - 2018-05-21
+
+* Use ora for build console message
+
 ## 0.2.1 - 2018-05-17
 
 * Log netlifySiteUrl when viewing CMS on localhost